Simplify ThreediDatabase
