+++ tests/test_checks_base.py
@@ -129,26 +129,6 @@ def test_null_check_with_null_value(session):
     assert invalid_rows[0].id == null_node.id
 
 
-def test_threedi_db_and_factories(threedi_db):
-    """Test to ensure that the threedi_db and factories use the same
-    session object."""
-    session = threedi_db.get_session()
-    factories.ManholeFactory()
-    q = session.query(models.Manhole)
-    assert q.count() == 1
-
-
-def test_run_spatial_function(session):
-    """Example how to use spatial functions.
-
-     Works on postgis and spatialite"""
-    factories.ConnectionNodeFactory()
-    from geoalchemy2 import func
-
-    q = session.query(func.ST_AsGeoJSON(models.ConnectionNode.the_geom))
-    q.first()
-
-
 def test_type_check(session):
     if session.bind.name == "postgresql":
         pytest.skip("type checks not working on postgres")
