+++ threedi_modelchecker/threedi_database.py
@@ -1,14 +1,7 @@
-# Copied from ThreeDiToolbox
-from builtins import object
-import copy
-
-from sqlalchemy.orm import sessionmaker
+from sqlalchemy.orm import sessionmaker, Session
 from sqlalchemy import create_engine
+from sqlalchemy.engine import Engine
 from sqlalchemy.event import listen
-from sqlalchemy.ext.declarative import declarative_base
-
-
-Base = declarative_base()
 
 
 def load_spatialite(con, connection_record):
@@ -41,76 +34,28 @@ def load_spatialite(con, connection_record):
     con.enable_load_extension(False)
 
 
-class ThreediDatabase(object):
-    def __init__(self, connection_settings, db_type="spatialite", echo=False):
-        """
+class ThreediDatabase:
 
-        :param connection_settings:
-        db_type (str choice): database type. 'sqlite' and 'postgresql' are
-                              supported
-        """
-        self.settings = connection_settings
-        self.db_type = db_type
-        self.echo = echo
+    def __init__(self, engine: Engine):
+        self.engine = engine
+        if self.engine.dialect.name == 'sqlite':
+            listen(engine, "connect", load_spatialite)
 
-        self._engine = None
-        self._base_metadata = None
+    @classmethod
+    def spatialite(cls, file, *args, **kwargs):
+        """Helper method to create a spatialite ThreediDatabase"""
+        engine = create_engine(f"sqlite:///{file}", *args, **kwargs)
+        return cls(engine)
 
-    @property
-    def engine(self):
-        return self.get_engine()
+    @classmethod
+    def postgis(cls, host, port, database, username, password, *args, **kwargs):
+        """Helper method to create a postgis ThreediDatabase"""
+        engine = create_engine(
+            f"postgresql://{username}:{password}@{host}:{port}/{database}", *args, **kwargs)
+        return cls(engine)
 
-    def get_engine(self, get_seperate_engine=False):
-
-        if self._engine is None or get_seperate_engine:
-            if self.db_type == "spatialite":
-                engine = create_engine(
-                    "sqlite:///{0}".format(self.settings["db_path"]), echo=self.echo
-                )
-                listen(engine, "connect", load_spatialite)
-                if get_seperate_engine:
-                    return engine
-                else:
-                    self._engine = engine
-
-            elif self.db_type == "postgres":
-                con = (
-                    "postgresql://{username}:{password}@{host}:"
-                    "{port}/{database}".format(**self.settings)
-                )
-
-                engine = create_engine(con, echo=self.echo)
-                if get_seperate_engine:
-                    return engine
-                else:
-                    self._engine = engine
-
-        return self._engine
-
-    def get_metadata(self, including_existing_tables=True, engine=None):
-
-        if including_existing_tables:
-            metadata = copy.deepcopy(Base.metadata)
-            if engine is None:
-                engine = self.engine
-
-            metadata.bind = engine
-            metadata.reflect(extend_existing=True)
-            return metadata
-        else:
-            if self._base_metadata is None:
-                self._base_metadata = copy.deepcopy(Base.metadata)
-            return self._base_metadata
-
-    def get_session(self):
-        return sessionmaker(bind=self.engine)()
+    def session(self, **kwargs) -> Session:
+        return sessionmaker(bind=self.engine, **kwargs)()
 
     def check_connection(self):
-        """Check if there a connection can be started with the database
-
-        :return: True if a connection can be established, otherwise raises an
-            appropriate error.
-        """
-        session = self.get_session()
-        r = session.execute("select 1")
-        return r.fetchone()
+        return self.engine.connect()
